# internal/models/auth.go
package models

type TokenPair struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}

type LoginResponse struct {
	User      *User     `json:"user"`
	TokenPair TokenPair `json:"tokens"`
}